Unify CSharpIdentifier
