+++ src/Compiler/Microsoft.AspNetCore.Mvc.Razor.Extensions.Version2_X/src/Microsoft.AspNetCore.Mvc.Razor.Extensions.Version2_X.csproj
@@ -14,9 +14,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\Microsoft.AspNetCore.Razor.Language\src\CodeGeneration\CodeWriterExtensions.cs">
-      <Link>Shared\CodeWriterExtensions.cs</Link>
-    </Compile>
+    <Compile Include="..\..\Microsoft.AspNetCore.Razor.Language\src\CodeGeneration\CodeWriterExtensions.cs" Link="Shared\CodeWriterExtensions.cs" />
+    <Compile Include="..\..\Microsoft.AspNetCore.Razor.Language\src\CSharpIdentifier.cs" Link="Shared\CSharpIdentifier.cs" />
   </ItemGroup>
 
   <ItemGroup>
