+++ src/Compiler/Microsoft.AspNetCore.Mvc.Razor.Extensions/src/MvcViewDocumentClassifierPass.cs
@@ -3,8 +3,6 @@
 
 #nullable disable
 
-using System;
-using Microsoft.AspNetCore.Razor;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Intermediate;
 
@@ -44,7 +42,7 @@ public class MvcViewDocumentClassifierPass : DocumentClassifierPassBase
             @namespace.Content = namespaceName;
         }
 
-        if (!TryComputeClassName(codeDocument, out var className))
+        if (!codeDocument.TryComputeClassName(out var className))
         {
             // It's possible for a Razor document to not have a file path.
             // Eg. When we try to generate code for an in memory document like default imports.
@@ -75,37 +73,5 @@ public class MvcViewDocumentClassifierPass : DocumentClassifierPassBase
         method.Modifiers.Add("async");
         method.Modifiers.Add("override");
         method.ReturnType = $"global::{typeof(System.Threading.Tasks.Task).FullName}";
-    }
-
-    private bool TryComputeClassName(RazorCodeDocument codeDocument, out string className)
-    {
-        var filePath = codeDocument.Source.RelativePath ?? codeDocument.Source.FilePath;
-        if (string.IsNullOrEmpty(filePath))
-        {
-            className = null;
-            return false;
-        }
-
-        className = GetClassNameFromPath(filePath);
-        return true;
-    }
-
-    private static string GetClassNameFromPath(string path)
-    {
-        var span = path.AsSpanOrDefault();
-
-        if (span.Length == 0)
-        {
-            return path;
-        }
-
-        const string cshtmlExtension = ".cshtml";
-
-        if (span.EndsWith(cshtmlExtension.AsSpan(), StringComparison.OrdinalIgnoreCase))
-        {
-            span = span[..^cshtmlExtension.Length];
-        }
-
-        return CSharpIdentifier.SanitizeIdentifier(span);
     }
 }
