+++ src/Compiler/Microsoft.AspNetCore.Mvc.Razor.Extensions/src/RazorPageDocumentClassifierPass.cs
@@ -5,7 +5,6 @@
 
 using System;
 using System.Diagnostics;
-using Microsoft.AspNetCore.Razor;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Extensions;
 using Microsoft.AspNetCore.Razor.Language.Intermediate;
@@ -69,7 +68,7 @@ public class RazorPageDocumentClassifierPass : DocumentClassifierPassBase
             @namespace.Content = namespaceName;
         }
 
-        if (!TryComputeClassName(codeDocument, out var className))
+        if (!codeDocument.TryComputeClassName(out var className))
         {
             // It's possible for a Razor document to not have a file path.
             // Eg. When we try to generate code for an in memory document like default imports.
@@ -168,37 +167,5 @@ public class RazorPageDocumentClassifierPass : DocumentClassifierPassBase
         {
             options.ParseLeadingDirectives = true;
         }
-    }
-
-    private bool TryComputeClassName(RazorCodeDocument codeDocument, out string className)
-    {
-        var filePath = codeDocument.Source.RelativePath ?? codeDocument.Source.FilePath;
-        if (string.IsNullOrEmpty(filePath))
-        {
-            className = null;
-            return false;
-        }
-
-        className = GetClassNameFromPath(filePath);
-        return true;
-    }
-
-    private static string GetClassNameFromPath(string path)
-    {
-        var span = path.AsSpanOrDefault();
-
-        if (span.Length == 0)
-        {
-            return path;
-        }
-
-        const string cshtmlExtension = ".cshtml";
-
-        if (span.EndsWith(cshtmlExtension.AsSpan(), StringComparison.OrdinalIgnoreCase))
-        {
-            span = span[..^cshtmlExtension.Length];
-        }
-
-        return CSharpIdentifier.SanitizeIdentifier(span);
     }
 }
